### 2019-11-03 #349 @LingshuHu
healthforum: A R package for scraping health forum discussion threads
Submitting Author Name: Lingshu Hu
Submitting Author Github Handle: @LingshuHu
Repository:  https://github.com/LingshuHu/healthforum
Version submitted: 0.1.0

Version accepted: TBD

---

```
Package: healthforum
Title: Scrape Patient Forum Data
Version: 0.1.0
    person("Lingshu", "Hu", ,
      email = "lingshu.hu@hotmail.com", role = c("aut", "cre"),
      comment = c(ORCID = "0000-0003-0304-882X")),
    person("Michael W.", "Kearney", ,
      email = "kearneymw@missouri.edu", role = c("ctb"),
      comment = c(ORCID = "0000-0002-0730-4694")))
Description: Scrape data from Patient Forum <https://patient.info/forums> by entering urls. It will return a data frame containing text, user names, like counts, reply counts, etc.
    rvest,
    xml2,
    tokenizers,
    tibble

```

	- [ ] workflow automataion

Data retrieval: this package is used to download data from *patient.info* forum.
Data extraction: the data obtained from Patient forum is unstructured, including text, date, behavior traces, etc.
Data munging: this package parses the unstructured data to dataframes.

Health communication scholars or researchers who are interested in user-generated health information.

To our best knowledge, there is no other R package designed to obtain data from *patient.info* forum.
