### 2019-11-07 #349 @maelle
Thanks for your submission @LingshuHu! We editors are discussing. We are specifically considering the privacy implications of this package. If you have any input on this topic, please let us know.
