+++ vignettes/basics.Rmd
@@ -110,3 +110,7 @@ counts <- count_medical_terms(text = medical_text)
 ## show results
 tibble::as_tibble(counts)
 ``` 
+
+
+### Disclaimer
+`healthforum` was developed to collect publicly available data from the website *patient.info*. The purpose of this package is to facilitate academic research. It is the final user's responsibility to store the data securely and obey all applicable local, state, and federal laws. 
