+++ README.Rmd
@@ -60,3 +60,5 @@ Preview the returned data frame
 tibble::as_tibble(gas)
 ```
 
+## Disclaimer
+`healthforum` was developed to collect publicly available data from the website *patient.info*. The purpose of this package is to facilitate academic research. It is the final user's responsibility to store the data securely and obey all applicable local, state, and federal laws. 
