### 2019-11-14 #349 @LingshuHu
Dear Maëlle,

Thank you for your feedback! In terms of privacy,
1) this package only collects publicly available data from the website. On this website, users need to fill out their names, nicknames, and other information. Users' real names are not visible, and users have control of showing information such as "about me," "my groups," "my replies," and "my discussions." This package only scrapes users' nicknames and other publicly visible information. By using users' nicknames, people cannot infer the URLs of users' profile pages and get their personal information.
2) We have added a responsibility disclaimer in README and vignette.
3) We are open to changes if reviewers suggest additional things to be done.

Please let us know if you have any other questions or comments. Thank you for your time and consideration!

Warm regards,
Lingshu
