### 2020-02-20 #349 @LingshuHu
Dear Maëlle,

I hope all is well. I would just like to check the status of our paper. Is there any further information that we could provide or any improvement that we can make? Thank you!

Regards,
Lingshu
