### 2020-02-21 #349 @noamross
Dear @LingshuHu, sorry this took a while to come back to.  In  the end we consulted with some bioethics and social media researchers and came up with a new Ethics and Privacy policy. You can preview it [here](https://5e50069c44685354d62e37aa--determined-hamilton-b43444.netlify.com/policies.html#ethics-data-privacy-and-human-subjects-research) and comment [here](https://github.com/ropensci/dev_guide/pull/251).
