### 2020-02-21 #349 @noamross
Based on this approach, here are my thoughts on the healthforum package: It clearly accesses generates personally identifiable and sensitive data.  Looking at all the relevant ToC/privacy/etc on the site, the normative expectations of privacy are quite ambiguous.  I think any research using the package would require a form of informed consent or a evaluation of such expectation, be it discussion with forum managers or a survey of users.  Since this is pretty much any use of the package, I think it makes sense to ask you,  the package authors, to do this so as to provide appropriate information to users.  This would mean contacting the site administrators to get their guidance on the use of the package and include that prominently in documentation.  It would also make sense to feature workflows in the vignette that removed personally identifying information.  For instance, you could show an example where you pull text and then generate analysis of frequently used terms, prominently noting that no user-level data is retained and this would be appropriate for publishable analysis.
