### 2020-02-24 #349 @mkearney
@noamross:

I believe the healthforum package only has access to screen names. So to clarify: is the policy of ropensci that **all 'user names' are considered to be personally identifiable data**?

The [GDPR defines personal data](https://gdpr-info.eu/art-4-gdpr/) as

> ‘personal data’ means any information relating to an identified or identifiable natural person (‘data subject’); an identifiable natural person is one who can be identified, directly or indirectly, in particular by reference to an identifier such as a name, an identification number, location data, an online identifier or to one or more factors specific to the physical, physiological, genetic, mental, economic, cultural or social identity of that natural person;

I think the above paragraph *could* be interpreted that way, but my understanding is that an **online identifier** counts as *personal data* **if it can be used to identify a natural person**. This makes sense if the user names *are* email addresses. Or if user names are used in combination with IP addresses or linked to actual names. But I worry what kind of effect the broader interpretation would have on research–or what it means to software like {rtweet} that was accepted in the past.
