### 2020-02-24 #349 @noamross
I responded over at https://github.com/ropensci/dev_guide/pull/251/ so as to consolidate conversation.
