### 2020-07-13 #349 @noamross
Hello @LingshuHu annd @mkearney, my apologies that this had fallen somewhat through the cracks somewhat without resolution.  We started review activities up a few weeks ago but I failed to pick up this conversation as it had spread across multiple repositories.

We adopted the policy that we discussed above and it is at https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research . Our take on **healthforum** is that, since the vast majority of research uses would require users to obtain a form of informed consent, the package authors should facilitate this by obtaining either blanket approval from patients.info, or, perhaps more realistically, appropriate contact info and a procedure for engaging with them and the user community and document this (e.g., "For informed consent procedures for using patients.info forum data, contact XXXX, community manager.")

Please let us know what your status is.  Again, sorry this took so long to pick up again.
