### 2020-08-03 #349 @LingshuHu
Dear @noamross ,

Thank you for updating! We really appreciate your work during this hard time! We think providing users with contact information would be a great idea. We have included it in README.Rmd and also remind users to contact their local IRBs to get more detailed information about privacy policies. We also created a package startup message containing this information. Whenever users library our package, they will see it.

Could you please let us know what we should do next? Would you suggest that we release a new version of the CRAN package first?
