+++ man/scrape_groups_by_initial_letter.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrape_data.R
+% Please edit documentation in R/deprecated_functions.R
 \name{scrape_groups_by_initial_letter}
 \alias{scrape_groups_by_initial_letter}
 \title{Scrape groups by initial letter}
