+++ README.Rmd
@@ -61,4 +61,4 @@ tibble::as_tibble(gas)
 ```
 
 ## Disclaimer
-`healthforum` was developed to collect publicly available data from the website *patient.info*. The purpose of this package is to facilitate academic research. It is the final user's responsibility to store the data securely and obey all applicable local, state, and federal laws. Users may want to contact *patient.info* to obtain a consent of collecting and using data. Their contact address is: Data Protection Officer, Fulford Grange, Micklefield Lane, Rawdon, Leeds, LS19 6BA. Email: privacy@emishealth.com.
+`healthforum` was developed to collect publicly available data from the website *patient.info*. The purpose of this package is to facilitate academic research. It is the final user's responsibility to store the data securely and obey all applicable local, state, and federal laws. For informed consent procedures for using patients.info forum data, please contact Data Protection Officer, Fulford Grange, Micklefield Lane, Rawdon, Leeds, LS19 6BA. Email: privacy@emishealth.com. 
