+++ NAMESPACE
@@ -7,7 +7,6 @@ S3method(count_medical_terms,list)
 export("%>%")
 export(count_medical_terms)
 export(scrape_groups_by_category)
-export(scrape_groups_by_initial_letter)
 export(scrape_one_group)
 export(scrape_one_post)
 export(scrape_user_posts)
