+++ R/healthforum-package.R
@@ -11,10 +11,4 @@ NULL
   packageStartupMessage(gsub("\\\n", " ", disclaimer))
 }
 
-disclaimer <- "`healthforum` was developed to collect publicly available data from
-the website *patient.info*. The purpose of this package is to facilitate academic
-research. It is the final user's responsibility to store the data securely and obey
-all applicable local, state, and federal laws. For informed consent procedures
-for using patients.info forum data, please contact Data Protection Officer,
-Fulford Grange, Micklefield Lane, Rawdon, Leeds, LS19 6BA.
-Email: privacy@emishealth.com."
+disclaimer <- "`healthforum` was developed to collect publicly available data from the website *patient.info*. The purpose of this package is to facilitate academic research. It is the final user's responsibility to store the data securely and obey all applicable local, state, and federal laws and ethical guidelines. For informed consent procedures for using patients.info forum data, please contact their Data Protection Officer at privacy@emishealth.com. Their address is Fulford Grange, Micklefield Lane, Rawdon, Leeds, LS19 6BA. You may also want to contact your local IRB to obtain information about privacy policy."
