### 2020-11-05 #349 @noamross
Dear @LingshuHu, my deep apologies that I did not respond to this previously.  Packages that had been underway during our "pause" fell through my alerts.

As I had written in my comment above, though, our opinion is that a disclaimer as included in the README isn't sufficient here, because, since the predominant uses of this package would require an informed consent procedure, doing the legwork of contacting the patients.info and obtaining either blanket approval or at least the relevant contact info and procedure, as I described above, is appropriate.

We realize this is a high standard and would be open to another option, such as including an IRB approval showing how certain uses can be approved without direct engagement with patients.info.
