### 2021-03-04 #349 @LingshuHu
Dear @noamross Thank you for reaching out to me! Yes, we want to continue to work on it. We plan to apply for an IRB review. But currently, I'm occupied by my dissertation and graduation. I will let you know when we get a chance to go through the IRB review.
